+++ otelcol/collector.go
@@ -26,7 +26,6 @@ import (
 	"go.opentelemetry.io/collector/extension"
 	"go.opentelemetry.io/collector/otelcol/internal/grpclog"
 	"go.opentelemetry.io/collector/processor"
-	"go.opentelemetry.io/collector/receiver"
 	"go.opentelemetry.io/collector/service"
 )
 
@@ -185,9 +184,12 @@ func (col *Collector) setupConfigurationComponents(ctx context.Context) error {
 	}
 
 	col.service, err = service.New(ctx, service.Settings{
-		BuildInfo:         col.set.BuildInfo,
-		CollectorConf:     conf,
-		Receivers:         receiver.NewBuilder(cfg.Receivers, factories.Receivers),
+		BuildInfo:     col.set.BuildInfo,
+		CollectorConf: conf,
+
+		ReceiversConfigs:   cfg.Receivers,
+		ReceiversFactories: factories.Receivers,
+
 		Processors:        processor.NewBuilder(cfg.Processors, factories.Processors),
 		Exporters:         exporter.NewBuilder(cfg.Exporters, factories.Exporters),
 		Connectors:        connector.NewBuilder(cfg.Connectors, factories.Connectors),
