+++ receiver/receivertest/nop_receiver_test.go
@@ -39,31 +39,3 @@ func TestNewNopFactory(t *testing.T) {
 	assert.NoError(t, logs.Start(context.Background(), componenttest.NewNopHost()))
 	assert.NoError(t, logs.Shutdown(context.Background()))
 }
-
-func TestNewNopBuilder(t *testing.T) {
-	builder := NewNopBuilder()
-	require.NotNil(t, builder)
-
-	factory := NewNopFactory()
-	cfg := factory.CreateDefaultConfig()
-	set := NewNopSettings()
-	set.ID = component.NewID(nopType)
-
-	traces, err := factory.CreateTracesReceiver(context.Background(), set, cfg, consumertest.NewNop())
-	require.NoError(t, err)
-	bTraces, err := builder.CreateTraces(context.Background(), set, consumertest.NewNop())
-	require.NoError(t, err)
-	assert.IsType(t, traces, bTraces)
-
-	metrics, err := factory.CreateMetricsReceiver(context.Background(), set, cfg, consumertest.NewNop())
-	require.NoError(t, err)
-	bMetrics, err := builder.CreateMetrics(context.Background(), set, consumertest.NewNop())
-	require.NoError(t, err)
-	assert.IsType(t, metrics, bMetrics)
-
-	logs, err := factory.CreateLogsReceiver(context.Background(), set, cfg, consumertest.NewNop())
-	require.NoError(t, err)
-	bLogs, err := builder.CreateLogs(context.Background(), set, consumertest.NewNop())
-	require.NoError(t, err)
-	assert.IsType(t, logs, bLogs)
-}
