+++ service/host.go
@@ -9,8 +9,8 @@ import (
 	"go.opentelemetry.io/collector/exporter"
 	"go.opentelemetry.io/collector/extension"
 	"go.opentelemetry.io/collector/processor"
-	"go.opentelemetry.io/collector/receiver"
 	"go.opentelemetry.io/collector/service/extensions"
+	"go.opentelemetry.io/collector/service/internal/builders"
 	"go.opentelemetry.io/collector/service/internal/graph"
 )
 
@@ -24,7 +24,7 @@ var _ component.Host = (*serviceHost)(nil)
 
 type serviceHost struct {
 	asyncErrorChannel chan error
-	receivers         *receiver.Builder
+	receivers         *builders.ReceiverBuilder
 	processors        *processor.Builder
 	exporters         *exporter.Builder
 	connectors        *connector.Builder
