+++ service/internal/graph/graph.go
@@ -30,7 +30,7 @@ import (
 	"go.opentelemetry.io/collector/exporter"
 	"go.opentelemetry.io/collector/internal/fanoutconsumer"
 	"go.opentelemetry.io/collector/processor"
-	"go.opentelemetry.io/collector/receiver"
+	"go.opentelemetry.io/collector/service/internal/builders"
 	"go.opentelemetry.io/collector/service/internal/capabilityconsumer"
 	"go.opentelemetry.io/collector/service/internal/status"
 	"go.opentelemetry.io/collector/service/pipelines"
@@ -41,7 +41,7 @@ type Settings struct {
 	Telemetry component.TelemetrySettings
 	BuildInfo component.BuildInfo
 
-	ReceiverBuilder  *receiver.Builder
+	ReceiverBuilder  *builders.ReceiverBuilder
 	ProcessorBuilder *processor.Builder
 	ExporterBuilder  *exporter.Builder
 	ConnectorBuilder *connector.Builder
